+++ .dir-locals.el
@@ -9,5 +9,6 @@
                                 ("test/\\([^/]+\\)-test.el" "tools/\\1.el")))
          (wal-project-build-default-cmd . "make tangle")
          (wal-project-install-default-cmd . ("make local"))
-         (wal-project-execute-default-cmd . ("make pacify" "make cold-boot" "make update-version"))))
+         (wal-project-execute-default-cmd . ("make pacify" "make cold-boot" "make update-version"))
+         (wal-project-install-reverse-mode . t)))
  (markdown-mode . ((eval . (auto-fill-mode t)))))